services/sam3_inference/requirements.txt
# SAM3 纯推理服务依赖
# 无 Label Studio 依赖

# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# SAM3
numpy>=1.26,<2.0
pillow>=9.0.0
opencv-python>=4.8.0
scikit-image>=0.19.0

# SAM3 模型

# 工具
python-dotenv==1.0.0
